# aspnetcore/src/IO.Swagger/Controllers/AttributesApi.cs
/*
 * Dilmah-Admin-BFF
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Collections.Generic;
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using IO.Swagger.Attributes;

using Microsoft.AspNetCore.Authorization;
using IO.Swagger.Models;

namespace IO.Swagger.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class AttributesApiController : ControllerBase
    { 
        /// <summary>
        /// get attribute by code
        /// </summary>
        /// <param name=""></param>
        /// <response code="200">success</response>
        [HttpGet]
        [Route("/attributes/{code}")]
        [ValidateModelState]
        [SwaggerOperation("AttributesCodeGet")]
        public virtual IActionResult AttributesCodeGet([FromRoute][Required] )
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200);

            throw new NotImplementedException();
        }

        /// <summary>
        /// Get attributes
        /// </summary>
        /// <response code="200">success</response>
        [HttpGet]
        [Route("/attributes")]
        [ValidateModelState]
        [SwaggerOperation("AttributesGet")]
        public virtual IActionResult AttributesGet()
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200);

            throw new NotImplementedException();
        }

        /// <summary>
        /// Create an attribute
        /// </summary>
        /// <response code="200">success</response>
        [HttpPost]
        [Route("/attributes")]
        [ValidateModelState]
        [SwaggerOperation("AttributesPost")]
        public virtual IActionResult AttributesPost()
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200);

            throw new NotImplementedException();
        }
    }
}