# aspnetcore/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.2",
  "info" : {
    "title" : "Dilmah-Admin-BFF",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "paths" : {
...
    "/options" : {
      "get" : {
        "tags" : [ "Options" ],
        "summary" : "Get all options",
        "responses" : {
          "200" : {
            "description" : "success",
            "content" : {
              "application/json" : {
                "examples" : {
                  "success" : {
                    "value" : {
                      "status" : true,
                      "data" : [ {
                        "id" : "",
                        "code" : "",
                        "name" : ""
                      } ]
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Options" ],
        "summary" : "create an option",
        "requestBody" : {
          "description" : "body",
          "content" : {
            "application/json" : {
              "examples" : {
                "body" : {
                  "value" : {
                    "code" : "",
                    "position" : 1,
                    "translations" : {
                      "de_CH" : {
                        "name" : "Bechergröße"
                      },
                      "en_US" : {
                        "name" : "Mug size"
                      }
                    },
                    "values" : [ {
                      "code" : "MUG_SIZE_S",
                      "translations" : {
                        "en_US" : {
                          "value" : "Small"
                        }
                      }
                    }, {
                      "code" : "MUG_SIZE_L",
                      "translations" : {
                        "en_US" : {
                          "value" : "Large"
                        }
                      }
                    } ]
                  }
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "success",
            "content" : {
              "application/json" : {
                "examples" : {
                  "success" : {
                    "value" : {
                      "status" : true,
                      "message" : "successfull"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/inventory/{code}" : {
      "get" : {
        "tags" : [ "Inventory" ],
        "summary" : "get inventory by code",
        "parameters" : [ {
          "name" : "code",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false
        } ],
        "responses" : {
          "200" : {
            "description" : "success",
            "content" : {
              "application/json" : {
                "examples" : {
                  "success" : {
                    "value" : {
                      "status" : true,
                      "data" : {
                        "name" : "",
                        "code" : "",
                        "inventory" : 10
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/attributes" : {
      "get" : {
        "tags" : [ "Attributes" ],
        "summary" : "Get attributes",
        "responses" : {
          "200" : {
            "description" : "success",
            "content" : {
              "application/json" : {
                "examples" : {
                  "success" : {
                    "value" : {
                      "status" : true,
                      "data" : [ {
                        "code" : "",
                        "name" : "",
                        "type" : ""
                      } ]
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Attributes" ],
        "summary" : "Create an attribute",
        "requestBody" : {
          "description" : "body",
          "content" : {
            "application/json" : {
              "examples" : {
                "body" : {
                  "value" : {
                    "id" : 1,
                    "code" : "mug_material",
                    "type" : "text",
                    "position" : 0,
                    "translations" : {
                      "de_CH" : {
                        "name" : "Becher Material"
                      }
                    }
                  }
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "success",
            "content" : {
              "application/json" : {
                "examples" : {
                  "success" : {
                    "value" : {
                      "status" : true,
                      "message" : "successfull"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/attributes/{code}" : {
      "get" : {
        "tags" : [ "Attributes" ],
        "summary" : "get attribute by code",
        "parameters" : [ {
          "name" : "code",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false
        } ],
        "responses" : {
          "200" : {
            "description" : "success",
            "content" : {
              "application/json" : {
                "examples" : {
                  "success" : {
                    "value" : {
                      "id" : 1,
                      "code" : "mug_material",
                      "type" : "text",
                      "position" : 0,
                      "translations" : {
                        "de_CH" : {
                          "id" : 1,
                          "locale" : "de_CH",
                          "name" : "Becher Material"
                        },
                        "en_US" : {
                          "id" : 2,
                          "locale" : "en_US",
                          "name" : "Mug material"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/options/{code}" : {
      "get" : {
        "tags" : [ "Options" ],
        "summary" : "get option by id",
        "parameters" : [ {
          "name" : "code",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false
        } ],
        "responses" : {
          "200" : {
            "description" : "success",
            "content" : {
              "application/json" : {
                "examples" : {
                  "success" : {
                    "value" : {
                      "status" : true,
                      "data" : {
                        "code" : "",
                        "position" : 1,
                        "translations" : {
                          "de_CH" : {
                            "name" : "Bechergröße"
                          },
                          "en_US" : {
                            "name" : "Mug size"
                          }
                        },
                        "values" : [ {
                          "code" : "MUG_SIZE_S",
                          "translations" : {
                            "en_US" : {
                              "value" : "Small"
                            }
                          }
                        }, {
                          "code" : "MUG_SIZE_L",
                          "translations" : {
                            "en_US" : {
                              "value" : "Large"
                            }
                          }
                        } ]
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}